Change structure of listenToDiscordBot & add deploy-commands.js
