+++ lib/index.js
@@ -2,3 +2,5 @@ export { listenToDiscordBot } from "./listenToDiscordBot.js";
 export { extractKeywordsFromText, initOpenAI } from "./extractKeywordsFromText.js";
 export { getOpportunitiesFromKeywords } from "./getOpportunitiesFromKeywords.js";
 export { getOpportunitiesInfoFromText } from "./getOpportunitiesInfoFromText.js";
+export { handleButtons } from "./handleButtons.js"
+export { jobSearch } from "./jobSearch.js"
